requirements.txt
streamlit>=1.29.0
boto3>=1.34.0
botocore>=1.34.0
Pillow>=10.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0